move tests to Deprecated9Tests for use of enhanced deprecation
